python-api/requirements.txt
fastapi>=0.110.0
uvicorn>=0.29.0
python-multipart>=0.0.9
python-dotenv>=1.0.0
# Reuse project libs
anthropic>=0.34.0
python-pptx>=0.6.21
Pillow
